+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - CLI option to set refresh interval (`--refresh-interval`)
+- Display peer country in torrent details
 
 ### Changed